requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
rank-bm25
redis
transformers
torch
numpy